### 2024-03-21 #6519 @editorialbot
[PRE REVIEW]: SubsetTools: A Python package to subset data to build and run ParFlow hydrologic models
**Submitting author:** @gartavanis (<a href="http://orcid.org/0009-0007-6015-7746">Georgios Artavanis</a>)
**Repository:** https://github.com/hydroframe/subsettools
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 1.0.3
