### 2024-03-21 #6519 @kthyng
Hi @gartavanis and thanks for your submission! I am looking for some specific items to make sure your submission fits our requirements at a high level before moving on to finding an editor or putting this on our waitlist if no relevant editors are available. I'll comment as I have a chance to go through them:
