### 2024-03-27 d552970
Update copyright (#74)
* add copyright

* update copyright

---------

Co-authored-by: George Artavanis <ga6@verde.princeton.edu>
