+++ README.md
@@ -23,4 +23,4 @@ Interested in contributing? Check out the contributing guidelines on the [Subset
 ## License
 `subsettools` is part of the [HydroFrame project](https://hydroframe.org/). It is licensed under the terms of the MIT license.
 
-Copyright ©2023 The Trustees of Princeton University and The Arizona Board of Regents on behalf of The University of Arizona.
+Copyright 2024 The Trustees of Princeton University and The Arizona Board of Regents on behalf of the University of Arizona, College of Science Hydrology & Atmospheric Sciences
