### 2024-03-28 #6519 @kthyng
@gartavanis Comments/questions:
* I see a "docs" dir — how do I see those docs, compiled somehow?
* I'm also looking for installation instructions which would presumably be there or in the readme (but I don't see in the readme)
