### 2024-04-01 #6519 @gartavanis
Hi Kristen,

the subsettools documentation is hosted on ReadTheDocs at

https://hydroframesubsettools.readthedocs.io/en/latest/

The documentation includes installation instructions under "Getting started".

Thanks,
George
