### 2024-04-01 #6519 @kthyng
@gartavanis Ok great yeah a good place is the little "About" section at the top right area on your github repo page. You should see a little gear to click on where you can add a docs link if you want it there.
