### 2024-04-16 748d444
clean up create_mask_solid
