### 2024-04-16 d66d5f5
factor out latlon checking code
