### 2024-04-17 aa1067f
remove unneeded argument from _indices_to_ij
