+++ tests/test_upstream_area.py
@@ -1,6 +1,5 @@
 import pytest
 import numpy as np
-import os
 from subsettools.upstream_area import delin_watershed
 
 # unit tests, multiple points, point outside grid, big area test, conus1 tests
@@ -26,7 +25,7 @@ from subsettools.upstream_area import delin_watershed
                                        (4054, 1914, 4058, 1917)),
                                       ([[39.8195, -75.3820]],
                                        "conus2",
-                                       np.load(os.path.abspath("tests/correct_output/upstream_mask_lower_delaware.npy")),
+                                       np.load("tests/correct_output/upstream_mask_lower_delaware.npy"),
                                        (3874, 1883, 4054, 2180)),
     ]
 )
