### 2024-04-26 2c9162d
cleanup and add test
