### 2024-04-26 7af7f9c
export upstream_area_to_ij function at the top level
