### 2024-04-29 10df1e8
add code to generate current year
