+++ docs/conf.py
@@ -6,8 +6,9 @@
 
 # -- Project information -----------------------------------------------------
 
+import datetime
 project = "subsettools"
-copyright = "2024 The Trustees of Princeton University and The Arizona Board of Regents on behalf of the University of Arizona, College of Science Hydrology & Atmospheric Sciences"
+copyright = f"{datetime.now().year} The Trustees of Princeton University and The Arizona Board of Regents on behalf of the University of Arizona, College of Science Hydrology & Atmospheric Sciences"
 author = "George Artavanis, Amanda Triplett"
 
 # -- General configuration ---------------------------------------------------
