### 2024-04-29 18b81e4
Upstream area subsetting, fixes #76 (#75)
* add copyright

* update copyright

* clean up create_mask_solid

* factor out latlon checking code

* remove unneeded argument from _indices_to_ij

* add upstream_area module

* add tests for upstream area subsetting

* add big upstream area test and reference output file

* remove unnecessary code

* cleanup and add test

* add tests

* format file with black

* cleanup code and format with black

* export upstream_area_to_ij function at the top level

* fix docstring

* exception fix

* Trigger Build

* add code to generate current year

* better v
[…]
