### 2024-05-02 #6519 @kellykochanski
Short on time unfortunately! I don't think I can commit to giving a review
up to JOSS's standard this month.

On Thu, May 2, 2024 at 5:23 AM Taher Chegini ***@***.***>
wrote:

> Hi @rreinecke <https://github.com/rreinecke>, @JannisHoch
> <https://github.com/JannisHoch>, and @kellykochanski
> <https://github.com/kellykochanski>! Would you like to review this
> submission to the Journal for Open Source Software? Our reviews are
> checklist-driven and openly conducted on GitHub over a timeline of 4–6
> weeks. Because the process is much more iterative and interactive than a
> traditional paper review, we would ask you to start within the next 2
> weeks. Here are reviewer guidelines for reference:
> joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> Thanks for your consideration.
>
