### 2024-05-03 26f5a1d
Fix upstream bug (#78)
* fix bug - return only bounding box mask and not entire conus2 mask

* fix tests to check bounding box masks

* replace with correct bounding box mask

* update changelog

* bump version

---------

Co-authored-by: George Artavanis <ga6@verde.princeton.edu>
