### 2024-05-03 54bec66
get latest hf_hydrodata version
