+++ pyproject.toml
@@ -12,7 +12,7 @@ numpy = ">=1.23.0"
 pytz = ">=2020.1"
 requests = ">=2.28.2"
 pftools = ">=1.3.10"
-hf-hydrodata = ">=1.0.8"
+hf-hydrodata = ">=1.1.9"
 
 [tool.poetry.dev-dependencies]
 
