### 2024-05-03 b897928
change huc_to_ij to return a mask in addition to the bounds
