### 2024-05-04 edd3cad
refactor latlon_to_ij
