### 2024-05-04 b994070
change latlon_to_ij to return right-bound inclusive bounds
