### 2024-05-04 2de53e3
modify latlon_to_ij to return mask of bounding box too
