+++ tests/test_subsettools.py
@@ -3,7 +3,6 @@ import pytest
 import subsettools as st
 from parflow.tools.io import read_pfb
 import numpy as np
-import os
 import hf_hydrodata
 from parflow import Run
 
@@ -33,12 +32,15 @@ def test_huc_to_ij_errors(huc_list, grid):
 
         
 @pytest.mark.parametrize(
-    "latlon_bounds, grid, result", [([[37.91, -91.43], [37.34, -90.63]], "conus1", (2285, 436, 2359, 496)),
-                                    ([[39.8379, -74.3791], [39.8379, -74.3791]], "conus2", (4057, 1915, 4058, 1916))]
+    "latlon_bounds, grid, correct_bounds", [([[37.91, -91.43], [37.34, -90.63]], "conus1", (2285, 436, 2359, 496)),
+                                            ([[39.8379, -74.3791], [39.8379, -74.3791]], "conus2", (4057, 1915, 4058, 1916))]
 )
-def test_latlon_to_ij(latlon_bounds, grid, result):
-    assert st.latlon_to_ij(latlon_bounds, grid) == result
-        
+def test_latlon_to_ij(latlon_bounds, grid, correct_bounds):
+    bounds, mask = st.latlon_to_ij(latlon_bounds, grid)
+    assert bounds == correct_bounds
+    imin, jmin, imax, jmax = bounds
+    assert np.array_equal(mask, np.ones((jmax - jmin, imax - imin), dtype=int))
+
 
 @pytest.mark.parametrize(
     "latlon_bounds, grid",
