### 2024-05-05 ee43041
refactor create_mask_solid
