### 2024-05-05 3964f40
modify tests to match new version of create_mask_solid
