### 2024-05-05 085a0a8
refactor create_mask_solid
