+++ pyproject.toml
@@ -13,13 +13,13 @@ pytz = ">=2020.1"
 requests = ">=2.28.2"
 pftools = ">=1.3.10"
 hf-hydrodata = ">=1.1.9"
+pylint = "^3.1.0"
 
 [tool.poetry.dev-dependencies]
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.0"
 pytest-cov = "^4.1.0"
-pylint = "^2.17.5"
 jupyter = "^1.0.0"
 matplotlib = "^3.7.2"
 black = "^23.9.1"
