### 2024-05-05 b2ab628
fix bug and change variable name
