### 2024-05-05 93eccc8
simplify comparison
