### 2024-05-06 f5e1311
replace get_ndarray calls
