+++ src/subsettools/subsettools.py
@@ -68,13 +68,14 @@ def huc_to_ij(huc_list, grid):
 
     huc_len = len(huc_list[0])
     huc_list = [int(huc) for huc in huc_list]
-    entry = hf_hydrodata.get_catalog_entry(
-        dataset="huc_mapping", grid=grid, file_type="tiff"
-    )
-    if entry is None:
-        raise ValueError(f"There is no HUC mapping entry for grid {grid}.")
-    conus_hucs = hf_hydrodata.gridded.get_ndarray(entry, level=str(huc_len))
-
+    try:
+        conus_hucs = hf_hydrodata.get_gridded_data(dataset="huc_mapping",
+                                                   grid=grid,
+                                                   file_type="tiff",
+                                                   level=str(huc_len),
+        )
+    except Exception as exc:
+        raise ValueError(f"Failed to get huc mapping data for the grid {grid}.") from exc
     huc_mask = np.isin(conus_hucs, huc_list).squeeze()
     indices_j, indices_i = np.where(huc_mask > 0)
     if indices_i.size == 0 or indices_j.size == 0:
@@ -291,17 +292,21 @@ def subset_static(
         raise TypeError("All variable names should be strings.")
     file_paths = {}
     for var in var_list:
-        entry = hf_hydrodata.get_catalog_entry(
-            dataset=dataset, file_type="pfb", period="static", variable=var
-        )
-        if entry is not None:
-            subset_data = hf_hydrodata.gridded.get_ndarray(entry, grid_bounds=ij_bounds)
+        try:
+            subset_data = hf_hydrodata.get_gridded_data(dataset=dataset,
+                                                        variable=var,
+                                                        file_type="pfb",
+                                                        period="static",
+                                                        grid_bounds=ij_bounds,
+            )
+        except ValueError:
+            print(f"Variable '{var}' not found in dataset '{dataset}'.")
+        else:
             file_path = os.path.join(write_dir, f"{var}.pfb")
             write_pfb(file_path, subset_data, dist=False)
             file_paths[var] = file_path
             print(f"Wrote {var}.pfb in specified directory.")
-        else:
-            print(f"{var} not found in dataset {dataset}")
+            
     return file_paths
 
 
@@ -346,21 +351,21 @@ def subset_press_init(ij_bounds, dataset, date, write_dir, time_zone="UTC"):
     if not isinstance(time_zone, str):
         raise TypeError("time_zone must be a string.")
     
-    entry = hf_hydrodata.get_catalog_entry(
-        dataset=dataset, file_type="pfb", variable="pressure_head", period="hourly"
-    )
-    if entry is None:
-        print(f"No pressure file found for in dataset {dataset}")
-        return None
-    
     new_date = get_UTC_time(date, time_zone)
-    print(f"UTC Date: {new_date}")
-        
+    print(f"UTC Date: {new_date}")        
     date_string = new_date.strftime("%Y.%m.%d:%H.%M.%S_UTC0")
 
-    subset_data = hf_hydrodata.gridded.get_ndarray(
-        entry, grid_bounds=ij_bounds, start_time=new_date
-    )
+    try:
+        subset_data = hf_hydrodata.get_gridded_data(dataset=dataset,
+                                                    variable="pressure_head",
+                                                    file_type="pfb",
+                                                    period="hourly",
+                                                    grid_bounds=ij_bounds,
+                                                    start_time=new_date,
+        )
+    except ValueError:
+        print(f"No pressure file found in dataset '{dataset}'.")
+        return None
 
     file_path = os.path.join(write_dir, f"{dataset}_{date_string}_press.pfb")
[…]
