### 2024-05-07 cf683bf
remove get_ndarray calls
