### 2024-05-07 cf88fa5
print ValueErrors from hf_hydrodata
