### 2024-05-07 69965b8
add mask check
