+++ src/subsettools/subset_utils.py
@@ -1,5 +1,4 @@
-"""This module contains helper functions for the subsetting module.
-"""
+"""This module contains helper functions for the subsetting module."""
 
 import os
 import shutil
