+++ src/subsettools/ref_yamls/__init__.py
@@ -1,2 +1,4 @@
-"""This subpackage contains the provided template runscripts. They can be accessed with the get_template_runscript() function.
+"""This subpackage contains the provided template runscripts. 
+
+The runscripts can be accessed with the get_template_runscript() function.
 """
