### 2024-05-07 82bd15d
fix docstring and add mask check
