+++ src/subsettools/subsettools.py
@@ -1,5 +1,4 @@
-"""This module contains subsetting functions for Parflow runs.
-"""
+"""This module contains subsetting functions for Parflow runs."""
 
 import os
 import shutil
@@ -25,6 +24,7 @@ from ._error_checking import (
     _validate_dir,
     _validate_grid_bounds,
     _validate_date,
+    _validate_mask,
 )
 
 
@@ -39,32 +39,43 @@ _HOURS_PER_FORCING_FILE = 24
 
 
 def huc_to_ij(huc_list, grid):
-    """Get the grid ij bounds of a bounding box that encompasses the HUC IDs provided in huc_list and a mask for that domain.
+    """Define a domain by a collection of HUCs.
 
-       All HUC IDs in huc_list must be the same length (HUCs of the same level). All HUCs 
-       should be adjacent. If a HUC is only partially covered by the provided grid, the grid 
-       bounds for the covered area will be returned.
+    The domain is defined by the grid ij bounds of a bounding box that
+    encompasses the HUCs in huc_list and a mask for that bounding box indicating
+    which cells are part of the HUCs.
+
+    All HUC IDs in huc_list must be the same length (HUCs of the same level).
+    All HUCs should be adjacent. If a HUC is only partially covered by the 
+    provided grid, the grid bounds for the covered area will be returned.
 
     Args:
         huc_list (list[str]): a list of USGS HUC IDs
-        grid (str): The spatial grid that the ij indices are calculated relative to and that the subset data will be returned on. Possible values: “conus1” or “conus2”
+        grid (str): The spatial grid that the ij indices are calculated relative
+            to and that the subset data will be returned on. Possible values: 
+            “conus1” or “conus2”
 
     Returns:
         A tuple (bounds, mask).
         
-        Bounds is a tuple of the form (imin, jmin, imax, jmax) representing the bounds in the conus grid of the area defined by the huc IDs in huc_list. imin, jmin, imax, jmax are the west, south, east and north sides of the box respectively and all i,j indices are calculated relative to the lower southwest corner of the domain.
+        Bounds is a tuple of the form (imin, jmin, imax, jmax) representing the
+        bounds in the conus grid of the area defined by the HUC IDs in huc_list.
+        imin, jmin, imax, jmax are the west, south, east and north sides of the
+        box respectively and all i,j indices are calculated relative to the
+        lower southwest corner of the domain.
 
-        Mask is a 2D numpy.ndarray that indicates which cells inside the bounding box are part of the selected HUC(s).
+        Mask is a 2D numpy.ndarray that indicates which cells inside the bounding
+        box are part of the selected HUC(s).
 
     Raises:
-        ValueError: If all HUC IDs are not the same length or if the area defined by the provided HUCs 
-            is not part of the given grid.
+        ValueError: If all HUC IDs are not the same length or if the area defined
+            by the provided HUCs is not part of the given grid.
 
     Example:
-
     .. code-block:: python
-
-        grid_bounds, mask = huc_to_ij(huc_list=["14080201", "14080202", "14080203"], grid="conus1")
+        grid_bounds, mask = huc_to_ij(
+            huc_list=["14080201", "14080202", "14080203"], grid="conus1"
+        )
     """
     _validate_huc_list(huc_list)
     _validate_grid(grid)
@@ -90,44 +101,57 @@ def huc_to_ij(huc_list, grid):
 
 
 def _indices_to_ij(indices_j, indices_i):
-    """Get the conus ij-bounds for the boundary defined by indices_j and indices_i.                                                  
+    """Get the grid ij bounds for the domain defined by indices_j and indices_i.
 
     Args:                                                                                                                           
[…]
