### 2024-05-08 5bc849c
fix docstrings and other pylint errors
