+++ src/subsettools/subset_utils.py
@@ -1,7 +1,6 @@
 """This module contains helper functions for the subsetting module."""
 
 import os
-import shutil
 from datetime import datetime, timedelta
 import pytz
 import numpy as np
@@ -9,22 +8,26 @@ from parflow.tools.io import read_clm
 
 _VEGM_COLUMNS = 25
 
+
 def subset_vegm(path, ij_bounds):
     """Read in vegm file and subset it according to ij_bounds.
 
-    Subset a national vegm file to a smaller domain based on ij_bounds that are provided relative to the grid that the national file is on. 
+    Subset a national vegm file to a smaller domain based on ij_bounds that are
+    provided relative to the grid that the national file is on.
 
     Args:
         path (str): path to read vegm file
-        ij_bounds (Tuple[int]): bounding box for subset. This should be given as i,j index values where 0,0 is the lower left hand 
-            corner of a domain. ij_bounds are given relative to whatever grid is being used for the subset. Use the latlon_to_ij function to 
-            determine ij indices from lat long values. 
+        ij_bounds (Tuple[int]): bounding box for subset. This should be given as
+            i,j index values where 0,0 is the lower left hand corner of a domain.
+            ij_bounds are given relative to whatever grid is being used for the
+            subset. Use the latlon_to_ij function to determine ij indices from
+            lat-lon values.
 
     Returns:
         ndarray:
             Subset vegm data.
     """
-    vegm_data = read_clm(path, type="vegm")    # returns (j,i,k)
+    vegm_data = read_clm(path, type="vegm")  # returns (j,i,k)
     vegm_data = np.transpose(vegm_data, (2, 0, 1))  # transpose to k,j,i
 
     imin, jmin, imax, jmax = ij_bounds
@@ -35,10 +38,10 @@ def subset_vegm(path, ij_bounds):
 
 def _reshape_ndarray_to_vegm_format(data):
     """Reshape ndarray returned by datacatalog to vegm format.
-    
+
     Args:
-        data (ndarray): raw subset vegm data (2d array)      
-    
+        data (ndarray): raw subset vegm data (2d array)
+
     Returns:
         Ndarray reshaped to vegm format.
     """
@@ -53,7 +56,8 @@ def _reshape_ndarray_to_vegm_format(data):
 def write_land_cover(land_cover_data, write_dir):
     """Write the land cover ndarray in vegm format.
 
-    Read in a gridded landcover dataset and write it out as a vegm file which is correctly formatted for CLM. 
+    Read in a gridded landcover dataset and write it out as a vegm file which is
+    correctly formatted for CLM.
 
     Args:
         land_cover_data (ndarray): formatted vegm data (2d array)
@@ -64,8 +68,8 @@ def write_land_cover(land_cover_data, write_dir):
             path to output vegm file.
     """
     heading = (
-        "x y lat lon sand clay color fractional coverage of grid, by vegetation class (Must/Should Add to "
-        "1.0) "
+        "x y lat lon sand clay color fractional coverage of grid, by "
+        "vegetation class (Must/Should Add to 1.0) "
     )
     vegm_col_names = (
         "",
@@ -116,41 +120,47 @@ def edit_drvclmin(
     vegp_name="drv_vegp.dat",
     vegm_name="drv_vegm.dat",
 ):
-    """Edit a template CLM driver for a new simulation. 
+    """Edit a template CLM driver for a new simulation.
 
-    Update the start and end dates, timezone, restart type and vegm and vegp input file names for a new simulation. 
+    Update the start and end dates, timezone, restart type and vegm and vegp
+    input file names for a new simulation.
 
     Args:
         file_path (str): clm driver file path
[…]
