+++ src/subsettools/subsettools.py
@@ -2,13 +2,12 @@
 
 import os
 import shutil
-import pathlib
 import subprocess
 from datetime import datetime, timedelta
-import threading 
+import threading
+import re
 import numpy as np
 import hf_hydrodata
-import re
 from parflow import Run
 from parflow.tools.io import read_pfb, write_pfb
 from .subset_utils import (
@@ -46,18 +45,18 @@ def huc_to_ij(huc_list, grid):
     which cells are part of the HUCs.
 
     All HUC IDs in huc_list must be the same length (HUCs of the same level).
-    All HUCs should be adjacent. If a HUC is only partially covered by the 
+    All HUCs should be adjacent. If a HUC is only partially covered by the
     provided grid, the grid bounds for the covered area will be returned.
 
     Args:
         huc_list (list[str]): a list of USGS HUC IDs
         grid (str): The spatial grid that the ij indices are calculated relative
-            to and that the subset data will be returned on. Possible values: 
+            to and that the subset data will be returned on. Possible values:
             “conus1” or “conus2”
 
     Returns:
         A tuple (bounds, mask).
-        
+
         Bounds is a tuple of the form (imin, jmin, imax, jmax) representing the
         bounds in the conus grid of the area defined by the HUC IDs in huc_list.
         imin, jmin, imax, jmax are the west, south, east and north sides of the
@@ -83,17 +82,22 @@ def huc_to_ij(huc_list, grid):
     huc_len = len(huc_list[0])
     huc_list = [int(huc) for huc in huc_list]
     try:
-        conus_hucs = hf_hydrodata.get_gridded_data(dataset="huc_mapping",
-                                                   grid=grid,
-                                                   file_type="tiff",
-                                                   level=str(huc_len),
+        conus_hucs = hf_hydrodata.get_gridded_data(
+            dataset="huc_mapping",
+            grid=grid,
+            file_type="tiff",
+            level=str(huc_len),
         )
     except Exception as exc:
-        raise ValueError(f"Failed to get huc mapping data for the grid {grid}.") from exc
+        raise ValueError(
+            f"Failed to get huc mapping data for the grid {grid}."
+        ) from exc
     huc_mask = np.isin(conus_hucs, huc_list).squeeze()
     indices_j, indices_i = np.where(huc_mask > 0)
     if indices_i.size == 0 or indices_j.size == 0:
-        raise ValueError(f"The area defined by the provided HUCs is not part of the {grid} grid.")  
+        raise ValueError(
+            f"The area defined by the provided HUCs is not part of the {grid} grid."
+        )
 
     bounds = _indices_to_ij(indices_j, indices_i)
     imin, jmin, imax, jmax = bounds
@@ -103,12 +107,12 @@ def huc_to_ij(huc_list, grid):
 def _indices_to_ij(indices_j, indices_i):
     """Get the grid ij bounds for the domain defined by indices_j and indices_i.
 
-    Args:                                                                                                                           
-        indices_j (numpy.ndarray): mask in the j direction for selected hucs                                                                
-        indices_i (numpy.ndarray): mask in the i direction for selected hucs                                                                   
- 
-    Returns:                                                                                                         
-        A tuple of the form (imin, jmin, imax, jmax) representing the bounds                                                                    
+    Args:
+        indices_j (numpy.ndarray): mask in the j direction for selected hucs
+        indices_i (numpy.ndarray): mask in the i direction for selected hucs
+
+    Returns:
+        A tuple of the form (imin, jmin, imax, jmax) representing the bounds
         in the grid defined by the two mask arrays indices_j and indices_i.
     """
[…]
