+++ src/subsettools/subsettools.py
@@ -71,6 +71,7 @@ def huc_to_ij(huc_list, grid):
             by the provided HUCs is not part of the given grid.
 
     Example:
+
     .. code-block:: python
         grid_bounds, mask = huc_to_ij(
             huc_list=["14080201", "14080202", "14080203"], grid="conus1"
