### 2024-05-08 e0e06c4
replace assertions with ValueErrors
