### 2024-05-09 324bcd7
deprecate huc_to_ij() and replace with define_huc_domain()
