+++ tests/test_multiple_hucs.py
@@ -17,7 +17,7 @@ def test_multiple_hucs(setup_dir_structure, remove_output_files):
         target_runscript,
     ) = setup_dir_structure(run_name)
 
-    huc_list = [
+    hucs = [
         "1302020701",
         "1302020702",
         "1408010604",
@@ -35,7 +35,7 @@ def test_multiple_hucs(setup_dir_structure, remove_output_files):
     Q = 1
     reference_run = st.get_template_runscript(grid, "transient", "solid", static_write_dir)
 
-    ij_bounds, mask = st.huc_to_ij(huc_list=huc_list, grid=grid)
+    ij_bounds, mask = st.define_huc_domain(hucs=hucs, grid=grid)
     st.create_mask_solid(mask=mask, grid=grid, write_dir=static_write_dir)
     st.subset_static(ij_bounds, dataset=var_ds, write_dir=static_write_dir)
     init_press_filepath = st.subset_press_init(
