+++ src/subsettools/subsettools.py
@@ -6,6 +6,7 @@ import subprocess
 from datetime import datetime, timedelta
 import threading
 import re
+import warnings
 import numpy as np
 import hf_hydrodata
 from parflow import Run
@@ -37,19 +38,19 @@ CONUS_Z_BOTTOM = 0
 _HOURS_PER_FORCING_FILE = 24
 
 
-def huc_to_ij(huc_list, grid):
+def define_huc_domain(hucs, grid):
     """Define a domain by a collection of HUCs.
 
     The domain is defined by the grid ij bounds of a bounding box that
-    encompasses the HUCs in huc_list and a mask for that bounding box indicating
-    which cells are part of the HUCs.
+    encompasses the HUCs in the list and a mask for that bounding box indicating
+    which cells in the bounding box are part of these HUCs.
 
-    All HUC IDs in huc_list must be the same length (HUCs of the same level).
+    All HUC IDs in hucs must be the same length (HUCs of the same level).
     All HUCs should be adjacent. If a HUC is only partially covered by the
     provided grid, the grid bounds for the covered area will be returned.
 
     Args:
-        huc_list (list[str]): a list of USGS HUC IDs
+        hucs (list[str]): a list of USGS HUC IDs
         grid (str): The spatial grid that the ij indices are calculated relative
             to and that the subset data will be returned on. Possible values:
             “conus1” or “conus2”
@@ -58,7 +59,7 @@ def huc_to_ij(huc_list, grid):
         A tuple (bounds, mask).
 
         Bounds is a tuple of the form (imin, jmin, imax, jmax) representing the
-        bounds in the conus grid of the area defined by the HUC IDs in huc_list.
+        bounds in the conus grid of the area defined by the HUC IDs in hucs.
         imin, jmin, imax, jmax are the west, south, east and north sides of the
         box respectively and all i,j indices are calculated relative to the
         lower southwest corner of the domain.
@@ -74,15 +75,15 @@ def huc_to_ij(huc_list, grid):
 
     .. code-block:: python
 
-        grid_bounds, mask = huc_to_ij(
-            huc_list=["14080201", "14080202", "14080203"], grid="conus1"
+        grid_bounds, mask = define_huc_domain(
+            hucs=["14080201", "14080202", "14080203"], grid="conus1"
         )
     """
-    _validate_huc_list(huc_list)
+    _validate_huc_list(hucs)
     _validate_grid(grid)
 
-    huc_len = len(huc_list[0])
-    huc_list = [int(huc) for huc in huc_list]
+    huc_len = len(hucs[0])
+    hucs = [int(huc) for huc in hucs]
     try:
         conus_hucs = hf_hydrodata.get_gridded_data(
             dataset="huc_mapping",
@@ -94,7 +95,7 @@ def huc_to_ij(huc_list, grid):
         raise ValueError(
             f"Failed to get huc mapping data for the grid {grid}."
         ) from exc
-    huc_mask = np.isin(conus_hucs, huc_list).squeeze()
+    huc_mask = np.isin(conus_hucs, hucs).squeeze()
     indices_j, indices_i = np.where(huc_mask > 0)
     if indices_i.size == 0 or indices_j.size == 0:
         raise ValueError(
@@ -104,6 +105,53 @@ def huc_to_ij(huc_list, grid):
     bounds = _indices_to_ij(indices_j, indices_i)
     imin, jmin, imax, jmax = bounds
     return bounds, huc_mask[jmin:jmax, imin:imax].astype(int)
+
+
+def huc_to_ij(huc_list, grid):
+    """This function is deprecated.
+
+    Use define_huc_domain() instead.
+    """
+    warnings.warn(
[…]
