+++ tests/test_conus1_upper_verde_spinup.py
@@ -17,7 +17,7 @@ def test_conus1_upper_verde_spinup(setup_dir_structure, remove_output_files):
         target_runscript,
     ) = setup_dir_structure(run_name)
 
-    huc_list = ["15060202"]
+    hucs = ["15060202"]
     start = "2005-10-01"
     grid = "conus1"
     run_ds = "conus1_baseline_mod"
@@ -26,7 +26,7 @@ def test_conus1_upper_verde_spinup(setup_dir_structure, remove_output_files):
     Q = 1
     reference_run = st.get_template_runscript(grid, "spinup", "solid", static_write_dir)
 
-    ij_bounds, mask = st.huc_to_ij(huc_list=huc_list, grid=grid)
+    ij_bounds, mask = st.define_huc_domain(hucs=hucs, grid=grid)
     st.create_mask_solid(mask=mask, grid=grid, write_dir=static_write_dir)
     st.subset_static(ij_bounds, dataset=var_ds, write_dir=static_write_dir)
     init_press_filepath = st.subset_press_init(
