### 2024-05-09 2b9a56d
deprecate latlon_to_ij() and replace with define_latlon_domain()
