### 2024-05-09 75b7f02
remove latlon_to_ij references in docstrings
