+++ src/subsettools/__init__.py
@@ -22,5 +22,5 @@ from subsettools.datasets import (
     get_template_runscript,
 )
 from subsettools.upstream_area import (
-    upstream_area_to_ij,
+    define_upstream_domain,
 )
