### 2024-05-09 7837c55
deprecate create_mask_solid() and replace with write_mask_solid()
