+++ tests/test_subsettools.py
@@ -53,13 +53,13 @@ def test_define_latlon_domain_errors(latlon_bounds, grid):
         st.define_latlon_domain(latlon_bounds, grid)
 
 
-def test_create_mask_solid(set_parflow_dir, tmp_path):
+def test_write_mask_solid(set_parflow_dir, tmp_path):
     hucs = ["15060202"]
     grid = "conus1"
     _, mask = st.define_huc_domain(hucs, grid)
     test_dir = tmp_path / "test"
     test_dir.mkdir()
-    st.create_mask_solid(mask, grid, test_dir)
+    st.write_mask_solid(mask, grid, test_dir)
     correct_output_dir = os.path.join(os.getcwd(), "tests", "correct_output", "conus1_upper_verde")
     match, mismatch, errors = filecmp.cmpfiles(test_dir, correct_output_dir, ["mask.pfb", "solidfile.pfsol"], shallow=False)
     assert len(match) == 2 and len(mismatch) == 0 and len(errors) == 0
