### 2024-05-11 b56c5cb
replace deprecated function call
