### 2024-05-11 02e998e
add encoding argument for open()
