### 2024-05-13 0ba9a91
Add documentation for mask vtk file in write_mask_solid()
