+++ pyproject.toml
@@ -28,6 +28,9 @@ sphinx-rtd-theme = "^1.3.0"
 sphinx = "6.2.1"
 myst-nb = {git = "https://github.com/executablebooks/MyST-NB.git"}
 
+[tool.pylint.messages_control]
+generated-members = ["domain", "domaininput", "indi_input"]
+
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
