### 2024-05-13 ef834c7
Remove backslash and format with black
