### 2024-05-14 491217b
Add decorator to deprecate old argument names
