### 2024-05-15 139dbc7
Replace deprecated arguments in tests
