+++ tests/test_conus1_upper_verde_spinup.py
@@ -46,8 +46,8 @@ def test_conus1_upper_verde_spinup(setup_dir_structure, remove_output_files):
         init_press=os.path.basename(init_press_filepath),
     )
     st.dist_run(
-        P=P,
-        Q=Q,
+        topo_p=P,
+        topo_q=Q,
         runscript_path=target_runscript,
         dist_clim_forcing=False,
     )
