### 2024-05-15 f2c35a9
Replace 'period' with 'temporal_resolution' in hf_hydrodata calls
