### 2024-05-15 393b64e
Include only active CONUS points in the mask
