+++ tests/test_subsettools.py
@@ -8,36 +8,60 @@ from parflow import Run
 
 
 @pytest.mark.parametrize(
-    "hucs, grid, result", [(["15060202"], "conus1", (375, 239, 487, 329)),
-                           (["140802"], "conus1", (572, 337, 797, 577)),
-                           (["14080201", "14080202", "14080203", "14080204", "14080205"], "conus1", (572, 337, 797, 577))]
+    "hucs, grid, result",
+    [
+        (["15060202"], "conus1", (375, 239, 487, 329)),
+        (["140802"], "conus1", (572, 337, 797, 577)),
+        (
+            ["14080201", "14080202", "14080203", "14080204", "14080205"],
+            "conus1",
+            (572, 337, 797, 577),
+        ),
+    ],
 )
 def test_define_huc_domain(hucs, grid, result):
     bounds, _ = st.define_huc_domain(hucs, grid)
     assert bounds == result
 
-    
+
 @pytest.mark.parametrize(
-    "hucs, grid", [(["01010001"], "conus1"),
-                   (["01010001"], "conus2"),
-                   (["03130003"], "conus1"),
-                   (["1710"], "conus1"),
-                   (["01010002", "01010001"], "conus1"),
-    ]
+    "hucs, grid",
+    [
+        (["01010001"], "conus1"),
+        (["01010001"], "conus2"),
+        (["03130003"], "conus1"),
+        (["1710"], "conus1"),
+        (["01010002", "01010001"], "conus1"),
+    ],
 )
 def test_define_huc_domain_errors(hucs, grid):
     """Check that a ValueError is raised if a HUC is not part of the grid."""
     with pytest.raises(ValueError) as e:
         st.define_huc_domain(hucs, grid)
 
-        
+
 @pytest.mark.parametrize(
     "latlon_bounds, grid, correct_bounds, correct_mask",
     [
-        ([[37.91, -91.43], [37.34, -90.63]], "conus1", (2285, 436, 2359, 496), np.ones((60, 74))),
-        ([[37.91, -91.43], [37.34, -90.63]], "conus2", (2684, 1403, 2758, 1461), np.ones((58, 74))),
-        ([[40.17, -124.27], [40.19, -124.25]], "conus2", (4, 2066, 7, 2069), np.array([[0, 0, 1], [0, 0, 1], [0, 1, 1]])),
-    ]
+        (
+            [[37.91, -91.43], [37.34, -90.63]],
+            "conus1",
+            (2285, 436, 2359, 496),
+            np.ones((60, 74)),
+        ),
+        (
+            [[37.91, -91.43], [37.34, -90.63]],
+            "conus2",
+            (2684, 1403, 2758, 1461),
+            np.ones((58, 74)),
+        ),
+        (
+            [[40.17, -124.27], [40.19, -124.25]],
+            "conus2",
+            (4, 2066, 7, 2069),
+            np.array([[0, 0, 1], [0, 0, 1], [0, 1, 1]]),
+        ),
+    ],
 )
 def test_define_latlon_domain(latlon_bounds, grid, correct_bounds, correct_mask):
     bounds, mask = st.define_latlon_domain(latlon_bounds, grid)
@@ -48,7 +72,11 @@ def test_define_latlon_domain(latlon_bounds, grid, correct_bounds, correct_mask)
 @pytest.mark.parametrize(
     "latlon_bounds, grid",
     [
-        pytest.param([[57.44, -107.33], [57.44, -107.33]], "conus2", id="lat/lon points outside the grid"),
+        pytest.param(
+            [[57.44, -107.33], [57.44, -107.33]],
+            "conus2",
+            id="lat/lon points outside the grid",
+        ),
[…]
