### 2024-05-16 7402414
Move domain-defining functions to domain.py module
