+++ src/subsettools/domain.py
@@ -1,0 +1,544 @@
+"""Functions to define a domain on the CONUS grids.
+
+A domain is defined by the grid ij bounds and a mask array. 
+
+The grid bounds is a tuple (imin, jmin, imax, jmax) that defines a bounding box
+on the CONUS grid that encompases the domain. Note that the origin (0, 0) of the
+CONUS grids is always the southwest corner. Thus, (imin, jmin) is the southwest
+corner of the bounding box, and (imax - 1, jmax - 1) the northeast corner (upper
+bounds are exclusive in keeping with Python convention).
+
+The mask is a 0/1 NumPy array which has the same shape as the bounding box and
+indicates which cells are part of the subset domain.
+
+SubsetTools provides functions to define domains in different ways.
+- A domain that is a collections of HUCs.
+- A box that is defined by two latitude-longitude points.
+- A domain that is the upstream area of a collection of outlets.
+"""
+
+
+import os
+import subprocess
+import warnings
+import numpy as np
+import hf_hydrodata
+from parflow.tools.io import write_pfb
+from ._error_checking import (
+    _validate_huc_list,
+    _validate_grid,
+    _validate_latlon_list,
+    _validate_dir,
+    _validate_mask,
+)
+
+
+CONUS_DX = 1000
+CONUS_DY = 1000
+CONUS1_DZ = 100
+CONUS1_Z_TOP = 500
+CONUS2_DZ = 200
+CONUS2_Z_TOP = 2000
+CONUS_Z_BOTTOM = 0
+
+
+def define_huc_domain(hucs, grid):
+    """Define a domain by a collection of HUCs.
+
+    The domain is defined by the grid ij bounds of a bounding box that
+    encompasses the HUCs in the list and a mask for that bounding box indicating
+    which cells in the bounding box are part of these HUCs.
+
+    All HUC IDs in hucs must be the same length (HUCs of the same level).
+    All HUCs should be adjacent. If a HUC is only partially covered by the
+    provided grid, the grid bounds for the covered area will be returned.
+
+    Args:
+        hucs (list[str]): a list of USGS HUC IDs
+        grid (str): The spatial grid that the ij indices are calculated relative
+            to and that the subset data will be returned on. Possible values:
+            “conus1” or “conus2”
+
+    Returns:
+        A tuple (bounds, mask).
+
+        Bounds is a tuple of the form (imin, jmin, imax, jmax) representing the
+        bounds in the conus grid of the area defined by the HUC IDs in hucs.
+        imin, jmin, imax, jmax are the west, south, east and north sides of the
+        box respectively and all i,j indices are calculated relative to the
+        lower southwest corner of the domain.
+
+        Mask is a 2D numpy.ndarray that indicates which cells inside the bounding
+        box are part of the selected HUC(s).
+
+    Raises:
+        ValueError: If the area defined by the provided HUCs is not part of the
+            given grid.
+
+    Example:
+
+    .. code-block:: python
+
+        grid_bounds, mask = define_huc_domain(
+            hucs=["14080201", "14080202", "14080203"], grid="conus1"
+        )
+    """
+    _validate_huc_list(hucs)
+    _validate_grid(grid)
+
+    huc_len = len(hucs[0])
+    hucs = [int(huc) for huc in hucs]
+    try:
+        conus_hucs = hf_hydrodata.get_gridded_data(
+            dataset="huc_mapping",
+            grid=grid,
+            file_type="tiff",
+            level=str(huc_len),
+        )
+    except Exception as exc:
+        raise ValueError(
+            f"Failed to get huc mapping data for the grid {grid}."
+        ) from exc
[…]
