### 2024-05-16 c30074a
Fix module docstring
