### 2024-05-16 2455642
Move functions that subset gridded files to subsetting.py
