+++ src/subsettools/subsettools.py
@@ -39,161 +39,6 @@ CONUS_Z_BOTTOM = 0
 _HOURS_PER_FORCING_FILE = 24
 
 
-def subset_static(
-    ij_bounds,
-    dataset,
-    write_dir,
-    var_list=(
-        "slope_x",
-        "slope_y",
-        "pf_indicator",
-        "mannings",
-        "pf_flowbarrier",
-        "pme",
-        "ss_pressure_head",
-    ),
-):
-    """Subset static input files from national datasets in HydroData.
-
-    The subset values will be written as ParFlow binary files (pfbs) in
-    write_dir. By default the following variables will be subset.
-        - Slope in the east/west direction (slope_x)
-        - Slope in the north/south direction (slope_y)
-        - Subsurface units indicator file (pf_indicator)
-        - Mannings roughness coefficients (mannings)
-        - Depth to bedrock (pf_flowbarrier)
-        - Long term average preciptation minus evaporation (i.e. recharge) (pme)
-        - Steady state pressure head used to initialize transient simulations
-          (ss_pressure_head)
-
-    Note that some datasets might not contain all 7 static input variables. In
-    that case, the subset_static function is going to get and save the data for
-    the variables supported by the dataset and print out a message for those
-    that are not.
-
-    Args:
-        ij_bounds (tuple[int]): bounding box for subset. This should be given as
-            i,j index values where 0,0 is the lower left hand corner of a domain.
-            ij_bounds are given relative to whatever grid is being used for the
-            subset.
-        dataset (str): static inputs dataset name from the HydroData catalog e.g.
-            "conus1_domain"
-        write_dir (str): directory where the subset files will be written
-        var_list (tuple[str]): tuple of variables to subset from the dataset.
-            By default all 7 variables above will be subset. The user can specify
-            as subset of these variables or list additional variables that are
-            available in their dataset of choice.
-
-    Returns:
-        A dictionary mapping the static variable names to the corresponding file
-        paths where the subset data were written.
-
-    Example:
-
-    .. code-block:: python
-
-        filepaths = subset_static(
-            ij_bounds=(375, 239, 487, 329),
-            dataset="conus1_domain",
-            write_dir="/path/to/your/chosen/directory",
-            var_list=("slope_x", "slope_y")
-        )
-    """
-    _validate_grid_bounds(ij_bounds)
-    if not isinstance(dataset, str):
-        raise TypeError("dataset name must be a string.")
-    _validate_dir(write_dir)
-    if not all(isinstance(var, str) for var in var_list):
-        raise TypeError("All variable names should be strings.")
-    file_paths = {}
-    for var in var_list:
-        try:
-            subset_data = hf_hydrodata.get_gridded_data(
-                dataset=dataset,
-                variable=var,
-                file_type="pfb",
-                temporal_resolution="static",
-                grid_bounds=ij_bounds,
-            )
-        except ValueError as err:
-            print(f"Variable '{var}' not found in dataset '{dataset}':", err)
-        else:
-            file_path = os.path.join(write_dir, f"{var}.pfb")
-            write_pfb(file_path, subset_data, dist=False)
-            file_paths[var] = file_path
-            print(f"Wrote {var}.pfb in specified directory.")
-
-    return file_paths
-
-
-def subset_press_init(ij_bounds, dataset, date, write_dir, time_zone="UTC"):
-    """Subset a pressure file from a national dataset in HydroData.
-
-    This function will select the pressure file for midnight on the date provided
-    and subset the selected pressure file to the ij_bounds provided. The subset
-    data will be written out as a ParFlow binary file (pfb) to be used as an
[…]
