### 2024-05-16 9ac3c3e
Add module for project-level constants
