+++ src/subsettools/_constants.py
@@ -1,0 +1,9 @@
+"""Project-level constants."""
+
+CONUS_DX = 1000
+CONUS_DY = 1000
+CONUS1_DZ = 100
+CONUS1_Z_TOP = 500
+CONUS2_DZ = 200
+CONUS2_Z_TOP = 2000
+CONUS_Z_BOTTOM = 0
