+++ src/subsettools/parflow_run.py
@@ -1,0 +1,383 @@
+"""Functions to get and customize a ParFlow runscript."""
+
+import os
+import shutil
+from importlib import resources
+from parflow import Run
+from parflow.tools.io import read_pfb
+from ._dev_utils import replace_kwargs
+from ._error_checking import (
+    _validate_grid,
+    _validate_dir,
+    _validate_grid_bounds,
+)
+from ._constants import (
+    CONUS_DX,
+    CONUS_DY,
+)
+
+def get_template_runscript(grid, mode, input_file_type, write_dir):
+    """Get a ParFlow template runscript.
+
+    The runscript is selected based on the grid, mode and input file type and
+    is copied to write_dir.
+
+    Args:
+        grid (str): The spatial grid that the ij indices are calculated relative
+            to and that the subset data will be returned on. Possible values:
+            “conus1” or “conus2”
+        mode (str): The type of simulation you would like to do. Possible values:
+            "spinup" (run ParFlow with a constant recharge forcing at the upper
+            boundary) and "transient" (coupled ParFlow-CLM run)
+        input_file_type (str): The type of domain you will run. Possible values:
+            "box" or "solid"
+        write_dir (str): directory where the template runscript file will be
+            copied
+
+    Returns:
+        A path to the template runscript.
+
+    Example:
+
+    .. code-block:: python
+
+        runscript_path = get_template_runscript(
+            grid="conus1",
+            mode="spinup",
+            input_file_type="solid",
+            write_dir="/path/to/your/chosen/directory"
+        )
+    """
+    _validate_grid(grid)
+    _validate_dir(write_dir)
+    if not isinstance(mode, str):
+        raise TypeError("mode must be a string")
+    if not isinstance(input_file_type, str):
+        raise TypeError("input_file_type must be a string")
+    if mode not in ["transient", "spinup"]:
+        raise ValueError("Supported modes are 'transient' and 'spinup'")
+    if input_file_type not in ["box", "solid"]:
+        raise ValueError("Supported input file types are 'box' and 'solid'")
+
+    filename = '_'.join([grid, mode, input_file_type]) + ".yaml"
+    with resources.as_file(
+        resources.files("subsettools.template_runscripts").joinpath(filename)
+    ) as f:
+        shutil.copy(f, write_dir)
+    return os.path.join(write_dir, filename)
+
+
+def edit_runscript_for_subset(
+    ij_bounds, runscript_path, write_dir=None, runname=None, forcing_dir=None
+):
+    """Modify a ParFlow run script for a new subdomain run.
+
+    This function is designed to start from a national ParFlow runscript template
+    and perform the following three modifications.
+        1. Modify the geometry to reflect the bounds of the desired ij_bounds
+           (i.e. the number of grid cells in the x and y direction and the upper
+           bounds of the geometry)
+        2. Update the runname to for the desired new run.
+        3. Update the location of the climate forcings for the new run.
+
+    If the runname is None and write_dir is the directory containing the
+    runscript file, the runscript file will be overwritten.
+
+    Args:
+        ij_bounds (tuple[int]): bounding box for subset. This should be given as
+            i,j index values where 0,0 is the lower left hand corner of a domain.
+            ij_bounds are given relative to whatever grid is being used for the
+            subset.
+        runscript_path (str): absolute path to the template parflow runscript
+            file
+        write_dir (str): directory where the new template file will be written.
+            If it is None, defaults to the directory containing the runscript.
+        runname (str): name for the new parflow run. If it is None, defaults to
+            the runscript's previous runname.
[…]
