+++ tests/test_parflow_run.py
@@ -7,6 +7,16 @@ import numpy as np
 from parflow import Run
 
 
+def test_get_ref_yaml_path(tmp_path):
+    test_dir = tmp_path / "test"
+    test_dir.mkdir()
+    runscript = st.get_template_runscript("conus1", "transient", "box", test_dir)
+    assert runscript == os.path.join(test_dir, "conus1_transient_box.yaml")
+    run = Run.from_definition(runscript)
+    assert run.ComputationalGrid.NX == 3342
+    assert run.GeomInput.domaininput.InputType == "Box"
+    
+
 def test_edit_runscript_for_subset_1(tmp_path):
     """Check the edited fiedls of the runscript file."""
     test_dir = tmp_path / "test"
