### 2024-05-16 e8f511c
Move config_clm and helper functions to clm.py
