+++ src/subsettools/clm.py
@@ -1,0 +1,250 @@
+"""Functions to configure CLM for coupled ParFlow-CLM simulations."""
+
+import os
+from datetime import timedelta
+import numpy as np
+import hf_hydrodata
+from .subset_utils import (
+    get_utc_time,
+    _reshape_ndarray_to_vegm_format,
+)
+from ._error_checking import (
+    _validate_dir,
+    _validate_grid_bounds,
+    _validate_date,
+)
+
+
+def config_clm(ij_bounds, start, end, dataset, write_dir, time_zone="UTC"):
+    """Modify template CLM driver files for a desired subdomain and run duration.
+
+    This function will obtain template clm driver files (specifically vegm, vep
+    and drv_clmin) from the existing national simulations on HydroData and modify
+    them to reflect the desired subdomain (indicated by the ij_bounds) and run
+    duration (indicated by the start and end dates). The modified files will be
+    written out to a user specified directory. These files are required if you
+    are going to run a ParFlow-CLM simulation.
+
+    Args:
+        ij_bounds (tuple[int]): bounding box for subset. This should be given as
+            i,j index values where 0,0 is the lower left hand corner of a domain.
+            ij_bounds are given relative to whatever grid is being used for the
+            subset.
+        start (str): start date (inclusive), in the form 'yyyy-mm-dd'
+        end (str): end date (exlusive), in the form 'yyyy-mm-dd'
+        dataset (str): the dataset that the files should be obtained from name
+            e.g. "conus1_baseline_mod"
+        write_dir (str): directory where the subset files will be written
+        timezone (str): timezone information for start and end dates.
+            Defaults to "UTC".
+
+    Returns:
+        A dictionary mapping the CLM file types ("vegp", "vegm", "drv_clm") to
+        the corresponging filepaths where the CLM files were written.
+
+    Example:
+
+    .. code-block:: python
+
+        filepaths = config_clm(
+            ij_bounds=(375, 239, 487, 329),
+            start="2005-10-01",
+            end="2006-10-01",
+            dataset="conus1_baseline_mod",
+            write_dir="/path/to/your/chosen/directory"
+        )
+    """
+    _validate_grid_bounds(ij_bounds)
+    _validate_date(start)
+    _validate_date(end)
+    if not isinstance(dataset, str):
+        raise TypeError("dataset name must be a string.")
+    _validate_dir(write_dir)
+    if not isinstance(time_zone, str):
+        raise TypeError("time_zone must be a string.")
+
+    # get the pfb version of the vegm file
+    file_type_list = ["vegp", "pfb", "drv_clm"]
+    file_paths = {}
+    for file_type in file_type_list:
+        if file_type == "vegp":
+            file_path = os.path.join(write_dir, "drv_vegp.dat")
+            try:
+                hf_hydrodata.get_raw_file(
+                    file_path,
+                    dataset=dataset,
+                    file_type=file_type,
+                    variable="clm_run",
+                    temporal_resolution="static",
+                )
+            except ValueError as err:
+                print(f"Failed to get {file_type} file for dataset '{dataset}':", err)
+            else:
+                file_paths[file_type] = file_path
+                print("copied vegp")
+        elif file_type == "pfb":
+            try:
+                subset_data = hf_hydrodata.get_gridded_data(
+                    dataset=dataset,
+                    file_type=file_type,
+                    variable="clm_run",
+                    temporal_resolution="static",
+                    grid_bounds=ij_bounds,
+                )
+            except ValueError as err:
[…]
