### 2024-05-16 31eac92
Reorganize helper functions in modules
