+++ src/subsettools/subsetting.py
@@ -15,7 +15,7 @@ import re
 import numpy as np
 import hf_hydrodata
 from parflow.tools.io import write_pfb
-from .subset_utils import (
+from ._common import (
     get_utc_time,
 )
 from ._error_checking import (
