### 2024-05-16 f5ae429
Update subsetting funcs (#79)
* get latest hf_hydrodata version

* change huc_to_ij to return a mask in addition to the bounds

* add latlon_to_ij tests

* refactor latlon_to_ij

* change latlon_to_ij to return right-bound inclusive bounds

* modify latlon_to_ij to return mask of bounding box too

* edit docstring

* add test

* refactor create_mask_solid

* modify tests to match new version of create_mask_solid

* refactor create_mask_solid

* add fixture setup_parflow_dir

* remove function

* update pylint version

* fix bug and change variable name

* simplify comparison

* replace get_nda
[…]
