+++ tests/test_subsettools.py
@@ -1,174 +1,0 @@
-import os
-import pytest
-import subsettools as st
-from parflow.tools.io import read_pfb
-import numpy as np
-import os
-import hf_hydrodata
-from parflow import Run
-
-
-@pytest.mark.parametrize(
-    "huc_list, grid, result", [(["15060202"], "conus1", (375, 239, 487, 329)),
-                               (["140802"], "conus1", (572, 337, 797, 577)),
-                               (["14080201", "14080202", "14080203", "14080204", "14080205"], "conus1", (572, 337, 797, 577))]
-)
-def test_huc_to_ij(huc_list, grid, result):
-    assert st.huc_to_ij(huc_list, grid) == result
-
-    
-@pytest.mark.parametrize(
-    "huc_list, grid", [(["01010001"], "conus1"),
-                       (["01010001"], "conus2"),
-                       (["03130003"], "conus1"),
-                       (["1710"], "conus1"),
-                       (["01010002", "01010001"], "conus1"),
-    ]
-)
-def test_huc_to_ij_errors(huc_list, grid):
-    """Check that a ValueError is raised if a HUC is not part of the grid."""
-    with pytest.raises(ValueError) as e:
-        st.huc_to_ij(huc_list, grid)
-    
-
-def test_forcing_timezones(tmp_path):
-    "Check if we get the correct forcing (temperature) in EST time."
-    utc = tmp_path / 'UTC_out'
-    utc.mkdir()
-    est = tmp_path / 'EST_out'
-    est.mkdir()
-    ij_bounds = (375, 239, 487, 329)
-    grid = 'conus1'
-    start = '2005-10-03' 
-    dataset = 'NLDAS2'
-    st.subset_forcing(ij_bounds=ij_bounds, grid=grid, start=start, end='2005-10-05', dataset=dataset, write_dir=utc, time_zone='UTC')
-    st.subset_forcing(ij_bounds=ij_bounds, grid=grid, start=start, end='2005-10-04', dataset=dataset, write_dir=est, time_zone='EST')
-    utc_temp1 = read_pfb(os.path.join(utc, "NLDAS.Temp.000001_to_000024.pfb"))
-    utc_temp2 = read_pfb(os.path.join(utc, "NLDAS.Temp.000025_to_000048.pfb"))
-    est_temp_correct = np.concatenate((utc_temp1[5:, :, :], utc_temp2[:5, :, :]), axis=0)
-    est_temp = read_pfb(os.path.join(est, "NLDAS.Temp.000001_to_000024.pfb"))
-    assert np.array_equal(est_temp_correct, est_temp)
-
-    
-def test_latlon_to_ij():
-    """Check that hf_hydrodata.grid.to_latlon and subsettools.latlon_to_ij are inverse to each other."""
-    latlon_points = hf_hydrodata.grid.to_latlon("conus1", *[375, 239, 487, 329])
-    latlon_points = [latlon_points[:2], latlon_points[2:]]
-    assert st.latlon_to_ij(latlon_points, "conus1") == (375, 239, 487, 329)
-
-    
-def test_edit_runscript_for_subset_1(tmp_path):
-    """Check the edited fiedls of the runscript file."""
-    test_dir = tmp_path / "test"
-    test_dir.mkdir()
-    runscript = st.get_template_runscript("conus1", "transient", "box", test_dir)
-    forcing_dir = tmp_path / "forcing"
-    forcing_dir.mkdir()
-    runscript = st.edit_runscript_for_subset((10, 10, 25, 25),
-                                                      runscript,
-                                                      runname="my_new_run",
-                                                      forcing_dir=str(forcing_dir)
-    )
[…]
