+++ src/subsettools/subset_utils.py
@@ -1,180 +1,0 @@
-"""This module contains helper functions for the subsetting module.
-"""
-
-import os
-import shutil
-from datetime import datetime, timedelta
-import pytz
-import numpy as np
-from parflow.tools.io import read_clm
-
-
-def subset_vegm(path, ij_bounds):
-    """Read in vegm file and subset it according to ij_bounds.
-
-    Subset a national vegm file to a smaller domain based on ij_bounds that are provided relative to the grid that the national file is on. 
-
-    Args:
-        path (str): path to read vegm file
-        ij_bounds (Tuple[int]): bounding box for subset. This should be given as i,j index values where 0,0 is the lower left hand 
-            corner of a domain. ij_bounds are given relative to whatever grid is being used for the subset. Use the latlon_to_ij function to 
-            determine ij indices from lat long values. 
-
-    Returns:
-        ndarray:
-            Subset vegm data.
-    """
-    vegm = read_clm(path, type="vegm")  # returns (j,i,k)
-    vegm = np.transpose(vegm, (2, 0, 1))  # transpose to k,j,i
-
-    imin, jmin, imax, jmax = ij_bounds
-    vegm = vegm[:, jmin:jmax, imin:imax]  # slicing on k,j,i
-
-    _, nj, ni = vegm.shape
-    indices = np.indices((nj, ni)) + 1
-    indices = indices[::-1, :, :]
-    vegm = np.vstack([indices, vegm])  # stack x,y indices on vegm
-
-    # transpose and reshape back into expected 2D vegm file format for the subset
-    vegm = vegm.transpose(1, 2, 0).reshape(-1, 25)
-
-    return vegm
-
-
-def write_land_cover(land_cover_data, write_dir):
-    """Write the land cover ndarray in vegm format.
-
-    Read in a gridded landcover dataset and write it out as a vegm file which is correctly formatted for CLM. 
-
-    Args:
-        land_cover_data (ndarray): formatted vegm data (2d array)
-        write_dir (str): path to output directory
-
-    Returns:
-        str:
-            path to output vegm file.
-    """
-    heading = (
-        "x y lat lon sand clay color fractional coverage of grid, by vegetation class (Must/Should Add to "
-        "1.0) "
-    )
-    vegm_col_names = (
-        "",
-        "",
-        "(Deg)",
-        "(Deg)",
-        "(%/100)",
-        "",
-        "index",
-        "1",
-        "2",
-        "3",
-        "4",
-        "5",
-        "6",
-        "7",
-        "8",
-        "9",
-        "10",
-        "11",
-        "12",
-        "13",
-        "14",
-        "15",
-        "16",
-        "17",
-        "18",
-    )
-    header = "\n".join([heading, " ".join(vegm_col_names)])
-    file_path = os.path.join(write_dir, "drv_vegm.dat")
-    np.savetxt(
-        fname=file_path,
-        X=land_cover_data,
-        delimiter=" ",
-        comments="",
-        header=header,
-        fmt=["%d"] * 2 + ["%.6f"] * 2 + ["%.2f"] * 2 + ["%d"] * 19,
-    )
-    return file_path
-
-
-def edit_drvclmin(
-    file_path,
-    start=None,
-    end=None,
-    time_zone="UTC",
-    startcode=2,
-    vegp_name="drv_vegp.dat",
-    vegm_name="drv_vegm.dat",
-):
-    """Edit a template CLM driver for a new simulation. 
-
-    Update the start and end dates, timezone, restart type and vegm and vegp input file names for a new simulation. 
-
-    Args:
-        file_path (str): clm driver file path
-        start (str): start date (inclusive), in the form 'yyyy-mm-dd'
[…]
