+++ src/subsettools/template_runscripts/__init__.py
@@ -1,0 +1,4 @@
+"""This subpackage contains the provided template runscripts. 
+
+The runscripts can be accessed with the get_template_runscript() function.
+"""

[…]
