+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "subsettools"
-version = "1.0.4"
+version = "1.0.5"
 description = "Subsetting tools and utilities for ParFlow"
 authors = ["George Artavanis, Amanda Triplett"]
 license = "MIT"
