### 2024-05-24 8a9990d
Factor out code to handle hf data requests
