### 2024-05-24 55d21f9
Refactor subset_static
