### 2024-05-24 4a2d194
Update subset_static documentation and examples
