### 2024-05-24 b848414
Refactor subset_static
