### 2024-05-24 7ae24cc
Refactor subset_forcing to improve exception handling
