### 2024-05-24 e583b64
If one subset_forcing thread fails, notify the other
