### 2024-05-24 976233c
Fix exception handling in subset_forcing
