### 2024-05-24 f6b4515
Add lock to synchronize access to shared dict
