### 2024-05-24 009062d
Warn users about new subset_static ValueError
