### 2024-05-31 d26c55a
Exception handling (#80)
* Factor out code to handle hf data requests

* Format with black

* Refactor subset_static

* Update subset_static documentation and examples

* Refactor subset_static

* Refactor subset_forcing to improve exception handling

* If one subset_forcing thread fails, notify the other

* Fix exception handling in subset_forcing

* Add lock to synchronize access to shared dict

* Warn users about new subset_static ValueError

* fix syntax error

* Replace get_gridded_data calls

* Replace get_gridded_data call for vegm

* Update changelog

* Bump version number

---------


[…]
