+++ src/subsettools/subsetting.py
@@ -12,11 +12,13 @@ import os
 from datetime import datetime, timedelta
 import threading
 import re
+import warnings
 import numpy as np
 import hf_hydrodata
 from parflow.tools.io import write_pfb
 from ._common import (
     get_utc_time,
+    get_hf_gridded_data,
 )
 from ._error_checking import (
     _validate_grid,
@@ -52,14 +54,15 @@ def subset_static(
         - Subsurface units indicator file (pf_indicator)
         - Mannings roughness coefficients (mannings)
         - Depth to bedrock (pf_flowbarrier)
-        - Long term average preciptation minus evaporation (i.e. recharge) (pme)
+        - Long term average precipitation minus evaporation (i.e. recharge) (pme)
         - Steady state pressure head used to initialize transient simulations
           (ss_pressure_head)
 
     Note that some datasets might not contain all 7 static input variables. In
-    that case, the subset_static function is going to get and save the data for
-    the variables supported by the dataset and print out a message for those
-    that are not.
+    that case, the subset_static function is going to raise a ValueError for any
+    variables that do not exist in the dataset. The default variable list
+    contains the necessary static variables for the CONUS2 grid. For CONUS1-based
+    datasets, "mannings" and "pf_flowbarrier" should be removed from the list.
 
     Args:
         ij_bounds (tuple[int]): bounding box for subset. This should be given as
@@ -71,7 +74,7 @@ def subset_static(
         write_dir (str): directory where the subset files will be written
         var_list (tuple[str]): tuple of variables to subset from the dataset.
             By default all 7 variables above will be subset. The user can specify
-            as subset of these variables or list additional variables that are
+            a subset of these variables or list additional variables that are
             available in their dataset of choice.
 
     Returns:
@@ -82,13 +85,34 @@ def subset_static(
 
     .. code-block:: python
 
+        # Subsetting static variables for a CONUS1 workflow
+        # We need to remove "pf_flowbarrier" and "mannings" from the list
         filepaths = subset_static(
             ij_bounds=(375, 239, 487, 329),
             dataset="conus1_domain",
             write_dir="/path/to/your/chosen/directory",
-            var_list=("slope_x", "slope_y")
+            var_list=("slope_x", "slope_y", "pf_indicator", "pme",
+                      "ss_pressure_head")
+        )
+
+        # Subsetting static variables for a CONUS2 workflow
+        # Note that we can use the default var_list here
+        filepaths = subset_static(
+            ij_bounds=(3701, 1544, 3792, 1633),
+            dataset="conus2_domain",
+            write_dir="/path/to/your/chosen/directory",
         )
     """
+    warnings.warn(
+        "Note that for subsettools versions >= 2.0.0, this function will raise "
+        "a ValueError if a variable in var_list is not supported in the "
+        "dataset. (In older versions, it just printed an error message and "
+        "continued executing normally). You can check in the HydroData "
+        "documentation which variables are contained in each dataset "
+        "(https://hf-hydrodata.readthedocs.io/en/latest/available_data.html).",
+        DeprecationWarning,
+        stacklevel=2,
+    )
     _validate_grid_bounds(ij_bounds)
     if not isinstance(dataset, str):
         raise TypeError("dataset name must be a string.")
@@ -96,22 +120,19 @@ def subset_static(
     if not all(isinstance(var, str) for var in var_list):
         raise TypeError("All variable names should be strings.")
     file_paths = {}
+    options = {
+        "dataset": dataset,
+        "file_type": "pfb",
+        "temporal_resolution": "static",
+        "grid_bounds": ij_bounds,
+    }
     for var in var_list:
-        try:
[…]
