+++ tests/test_conus1_upper_verde.py
@@ -123,29 +123,4 @@ def test_conus1_upper_verde(setup_run, remove_output_files, P, Q):
     run = Run.from_definition(target_runscript)
     run.TimingInfo.StopTime = 10
     run.run(working_directory=pf_out_dir)
-
-    test_vars = ["perm_x", "perm_y", "perm_z"]
-    for var in test_vars:
-        filename = f"{run_name}.out.{var}.pfb"
-        assert pf_test_file(
-            os.path.join(pf_out_dir, filename),
-            os.path.join(correct_output_dir, filename),
-            "Max difference in perm_x",
-        )
-
-    for i in range(5):
-        filename = f"{run_name}.out.press.0000{i}.pfb"
-        assert pf_test_file(
-            os.path.join(pf_out_dir, filename),
-            os.path.join(correct_output_dir, filename),
-            f"Max difference in Pressure for timestep {i}",
-        )
-
-        filename = f"{run_name}.out.satur.0000{i}.pfb"
-        assert pf_test_file(
-            os.path.join(pf_out_dir, filename),
-            os.path.join(correct_output_dir, filename),
-            f"Max difference in Saturation for timestep {i}",
-        )
-
     remove_output_files(pf_out_dir)
