### 2024-06-09 6ded22f
Add hf.register_api_pin call to conftest
