### 2024-06-09 1513af9
Update tests (#82)
* Update ParFlow integration tests

* Add hf.register_api_pin call to conftest

---------

Co-authored-by: gartavanis <gartavanis@users.noreply.github.com>
