### 2024-06-13 4c03663
adds bug fix for mask value checking
