### 2024-06-14 #6752 @dvalters
Hi @cheginit The software meets all the requirements to be accepted by JOSS and I suggest it be accepted as is.

I could not get every single test to run on my machine, however I suspect that is down to requiring extra data sources for certain tests that I don't have. The software clearly has a fairly comprehensive suite of tests and I'm happy for it to pass that criteria without me needing to verify every single unit test. (which I would not have time to do)

> [...] if the submitting authors have followed best practices (have documentation, tests, continuous integration, and a license) then their review should be rapid.

A nice piece of software @JannisHoch :)
