### 2024-06-14 #6752 @cheginit
@dvalters Thanks for taking the time to review the package. Usually, `pytest` tests can be run with a single `pytest` command invocation. @gartavanis, I noticed that in your contribution file, there's no explicit guidance on running the tests. Also, there's no contribution instruction on the website. Please make sure to add them.

@JannisHoch Thanks for the update.
