### 2024-06-14 #6752 @dvalters
@cheginit yes I was able to invoke the test suite using `pytest` - I just meant that some of the tests failed, but it looked like an input data issue (so specific tests failed as presumably they rely on an external data source?)
