### 2024-06-14 #6752 @cheginit
@dvalters I see. In that case, I'd appreciate it if you open an issue in their repo and report the failing tests, i.e., the log of running `pytest`, so the authors can address the issue or justify them.
