### 2024-06-21 #6752 @cheginit
@JannisHoch In this context, original means that the data and results are produced for/by the software. For example, original data means that new data has been prepared for running the software and original results mean that the software produced new results that must be verifiable by following the instructions provided by the authors.
