### 2024-06-21 hydroframe/subsettools#86 @JannisHoch
JOSS/review/code
Overall, the code is well documented and the readthedoc documentation is exhaustive.

To finalise the JOSS review, some minor comments need to be adressed still.

- [ ] In https://hydroframesubsettools.readthedocs.io/en/latest/example_notebooks/example_notebooks.html, it would be great if it is mentioned that the notebooks can be run via Binder, and that a Princeton HydroGEN PIN is necessary. Great btw that acquiring the PIN is very straightforward!
- [ ] With respect to the notebooks available on https://hydroframesubsettools.readthedocs.io/en/latest/example_notebooks/example_notebooks.html: it would be good to mention how to run the notebooks locally. Not a big deal if one is used to work with these notebooks, but since your packages is intended for less experiences users, more information would be useful.
- [ ] It is not possible to run `pytest` in the `test` folder due to the following error message: `FileNotFoundError: [Errno 2] No such file or directory: 'tests/correct_output/upstream_mask_lower_delaware.npy'`. Would it be possible to make the files necessary for the tests available?

Many thanks!
