### 2024-06-21 4019e70
adds simple write_solid_mask tests
