### 2024-06-21 72a46f4
Add guidelines for running tests
