### 2024-06-21 #6752 @gartavanis
Thanks for your review @dvalters! I'm happy to look at your pytest log, and address any issues there.

@cheginit Thanks for your suggestions. I've added instructions for running tests on the contributing file. Also, I fixed the link to the contributing instructions on the README.
