### 2024-06-24 d0f57ec
adds mask dimension check
