### 2024-06-24 5e2bc6c
adds file assertion and updates error messages
