### 2024-06-24 72abdea
adds a couple conus2 conditions to test_define_huc_domain
