### 2024-06-25 bd6ab7c
adds test cases for define_huc_domain
