### 2024-06-26 e0cc465
updates define huc domain error testing
