+++ src/subsettools/_constants.py
@@ -7,4 +7,3 @@ CONUS1_Z_TOP = 500
 CONUS2_DZ = 200
 CONUS2_Z_TOP = 2000
 CONUS_Z_BOTTOM = 0
-#comment
