+++ paper/paper.md
@@ -141,6 +141,10 @@ runscript_path = st.edit_runscript_for_subset(
 
 The `SubsetTools` package also provides functions to customize the template runscript, for example by specifying the desired subset domain to match the subset inputs, modifying the file paths of the model input files, and changing the processor topology for the ParFlow run. Once the customized Parflow runscript is ready, the user can launch a ParFlow simulation using the [pftools](https://pypi.org/project/pftools/) package utilities. 
 
+# Future Work
+The current version of `SubsetTools` is a novel package to create and run ParFlow watershed models. However, this package remains under active development to add new features. For example, we plan to add functions to assist users in restarting a ParFlow run and heuristically estimate processor topology. Further, we plan to expand the methods by which a user can define a domain such as by the area upstream of a point or with a shapefile. Finally, we currenly only create watershed models within CONUS. The package can easily be expanded to cover other areas of the world if the appropriate model input files are provided to be hosted in `Hydrodata`. 
+
+
 # Acknowledgements
 
 This research has been supported by the U.S. Department of Energy Office of Science IDEAS-Watersheds (DE-AC02-05CH11231), the US National Science Foundation Office of Advanced Cyberinfrastructure HydroFrame projects (OAC- 2054506 and OAC-1835855) and the US National Science Foundation HydroGEN project (NSF C-A-2134892).
