### 2024-06-28 353aa47
removes allowed values variable
