+++ src/subsettools/_error_checking.py
@@ -75,6 +75,5 @@ def _validate_mask(mask):
     if len(mask.shape) != 2:
         raise ValueError("mask must be 2D numpy array")
     unique_values = np.unique(mask)
-    allowed_values = [0,1]
-    if len(unique_values) > 2 or not all(v in allowed_values for v in unique_values):
+    if len(unique_values) > 2 or not all(v in [0,1] for v in unique_values):
         raise ValueError("mask elements must be 0 or 1.")
