### 2024-06-28 871535d
adds testing conditions for define huc domain tests
