### 2024-06-28 b9c2f11
adds testing conditions for define lat/lon domain tests
