### 2024-06-28 1e0ed58
Update example_notebooks.md
