### 2024-06-28 f2ac67e
Update example_notebooks.md
