### 2024-06-28 7173957
Update example_notebooks.md
