+++ docs/example_notebooks/example_notebooks.md
@@ -12,9 +12,9 @@ cd subsettools/docs/example_notebooks
 
 There are three ways to run the notebooks, as explained in more detail in the [Getting started](https://hydroframesubsettools.readthedocs.io/en/latest/getting_started.html) section.
 
-Each notebook can be run in `Binder`. Click the `launch binder` button at the top of each example notebook.
+Each notebook can be run in Binder. Click the `launch binder` button at the top of each example notebook.
 
-The notebooks can also be run in `Docker`. Follow the instructions in [Getting started](https://hydroframesubsettools.readthedocs.io/en/latest/getting_started.html) to get the `subsettools` Docker container and run the notebooks.
+The notebooks can also be run in Docker. Follow the instructions in [Getting started](https://hydroframesubsettools.readthedocs.io/en/latest/getting_started.html) to get the `subsettools` Docker container and run the notebooks.
 
 Finally, the notebooks can be run on your local machine. Clone the `subsettools` repo or download an individual notebook from GitHub, and run the notebooks with a kernel that has the package pip-installed. Note that if you're running locally, in order to complete some workflows, you will also need to have ParFlow installed. Please refer to the [HydroFrame](https://hydroframe.org/) for help getting started with ParFlow.
 
