### 2024-06-28 a24d5af
Update example_notebooks.md (#90)
* Update example_notebooks.md

* Update example_notebooks.md

* Update example_notebooks.md
