### 2024-07-01 12f70f2
corrects latlon domain test params
