### 2024-07-01 807c94c
Monkeypatch hf.get_gridded_data for tests
