### 2024-07-01 hydroframe/subsettools#86 @gartavanis
Thank you, @JannisHoch, for your comments. We will address them below:

1 and 2: We have updated the documentation on ReadTheDocs to provide instructions for running the notebooks on Binder, Docker and on one's local machine. Glad to hear that the HydroGEN PIN was easy to get and use!

3: Are you running `pytest` from the root of the subsettools directory (i.e. above the "tests" directory)? I can see the delaware file exists on the repo (https://github.com/hydroframe/subsettools/tree/main/tests/correct_output), so I wonder if it's an issue with the directory from which `pytest` is run.
