### 2024-07-02 hydroframe/subsettools#86 @JannisHoch
Hi @gartavanis,

thanks for updating the RTD documentation.

Regarding `pytest`, this is what I get when running `pytest` in `subsettools/tests` in branch `main`:

```
======================================================================================================= ERRORS =======================================================================================================
_______________________________________________________________________________________ ERROR collecting tests/test_domain.py ________________________________________________________________________________________
test_domain.py:127: in <module>
    np.load("tests/correct_output/upstream_mask_lower_delaware.npy"),
../../../../opt/anaconda3/envs/subsettools/lib/python3.11/site-packages/numpy/lib/_npyio_impl.py:455: in load
    fid = stack.enter_context(open(os.fspath(file), "rb"))
E   FileNotFoundError: [Errno 2] No such file or directory: 'tests/correct_output/upstream_mask_lower_delaware.npy'
============================================================================================== short test summary info ===============================================================================================
ERROR test_domain.py - FileNotFoundError: [Errno 2] No such file or directory: 'tests/correct_output/upstream_mask_lower_delaware.npy'
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 1 error during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
================================================================================================== 1 error in 3.13s ==================================================================================================
```
