### 2024-07-02 hydroframe/subsettools#86 @JannisHoch
Indeed, if I execute `pytest` in the main directory, executes tests but fails on 16 tests (see attachment)?

[subsettools_pytest.txt](https://github.com/user-attachments/files/16062999/subsettools_pytest.txt)
