### 2024-07-02 78a1322
Add pseudocode for mock hf call
