### 2024-07-02 7c0ee3e
Add more options to mock_get_data
