### 2024-07-03 #6752 @cheginit
👋🏼 @dvalters Will you be able to open the issue that I mentioned in my earlier comment?

Here's the comment:
> I see. In that case, I'd appreciate it if you open an issue in their repo and report the failing tests, i.e., the log of running `pytest`, so the authors can address the issue or justify them.
