### 2024-07-03 a1c9c86
adds subset press init test
