### 2024-07-03 723e953
adds mocking and test for subset_forcing
