### 2024-07-04 hydroframe/subsettools#86 @JannisHoch
Ah okay great, thanks for the explanation.
Would be good to mention somewhere how to provide the PIN and mail for the tests. Since running the tests seems to be rather elaborate, I could imagine there is a case for a dedicated section about it?
