### 2024-07-08 ddb8261
adds subset forcing file name testing
