+++ tests/test_subsetting.py
@@ -53,13 +53,14 @@ def mock_hf(monkeypatch):
     
     monkeypatch.setattr(hf, "get_gridded_data", mock_get_data)
 
+
 @pytest.fixture
 def mock_get_paths(monkeypatch):
 
     def mock_hf_get_paths(options):
         dataset = options.get("dataset")
         variable = options.get("variable")
-        file_path = [f"/folder/{dataset}.{variable}.000001_to_000024.pfb"]
+        file_path = [f"/folder/{dataset}.{variable}.000000_to_000000.pfb"]
         return file_path
 
     monkeypatch.setattr(hf, "get_paths", mock_hf_get_paths)
@@ -80,6 +81,7 @@ def test_subset_static(tmp_path, mock_hf):
     assert os.path.isfile(paths["var2"])
     data2 = read_pfb(paths["var2"])
     assert np.array_equal(data2, np.ones((_DUMMY_NZ, 20, 10)))
+
 
 @pytest.mark.parametrize(
     "time_zone",
@@ -102,6 +104,7 @@ def test_subset_press_init_function(time_zone, tmp_path, mock_hf):
     data = read_pfb(file_path)
     assert np.array_equal(data, np.ones((_DUMMY_NZ, 20, 10)))
 
+
 @pytest.mark.parametrize(
     "time_zone",
     [
@@ -123,13 +126,30 @@ def test_subset_forcing(time_zone, tmp_path, mock_hf, mock_get_paths):
         time_zone=time_zone
     )
     assert os.path.isfile(paths["var1"][0])
+    assert os.path.isfile(paths["var1"][1])
     assert np.array_equal(read_pfb(paths["var1"][0]), np.ones((24, 20, 10)))
     assert np.array_equal(read_pfb(paths["var1"][1]), np.ones((24, 20, 10)))
     assert np.array_equal(read_pfb(paths["var2"][0]), np.ones((24, 20, 10)))
     assert np.array_equal(read_pfb(paths["var2"][1]), np.ones((24, 20, 10)))
-    
 
-    
+
+def test_subset_forcing_files(tmp_path, mock_hf, mock_get_paths):
+    test_dir = tmp_path / "test_forcing_files"
+    test_dir.mkdir()
+    paths = st.subset_forcing(
+        ij_bounds=(0,0,10,20),
+        grid="conus1",
+        start="2005-10-02", 
+        end="2005-10-05", 
+        dataset="my_ds",
+        forcing_vars=(["var1"]),
+        write_dir=test_dir,
+        )
+    print(type(os.path.basename(paths["var1"][0])), os.path.basename(paths["var1"][0]))
+    assert os.path.basename(paths["var1"][0]) == "my_ds.var1.000001_to_000024.pfb"
+    assert os.path.basename(paths["var1"][1]) == "my_ds.var1.000025_to_000048.pfb"
+    assert os.path.basename(paths["var1"][2]) == "my_ds.var1.000049_to_000072.pfb"
+
 
 def test_forcing_timezones(tmp_path):
     "Check if we get the correct forcing (temperature) in EST time."
