### 2024-07-09 5159dd9
Remove test requiring 16 cores as redundant
