### 2024-07-09 d449756
Add instructions for running test suite
