### 2024-07-09 85d9369
Update docs (#92)
* Remove test requiring 16 cores as redundant

* Add instructions for running test suite

* Update CONTRIBUTING.md

---------

Co-authored-by: gartavanis <gartavanis@users.noreply.github.com>
