### 2024-07-09 hydroframe/subsettools#86 @gartavanis
Thanks for the suggestion, @JannisHoch!
I've added a "Running the test suite" section to the Contributing page in the documentation with instructions.
I've also removed the test that requires 16 cores so that should no longer be an issue.
