### 2024-07-09 7002885
Cleanup and format code
