+++ tests/test_subsetting.py
@@ -4,6 +4,7 @@ import numpy as np
 import subsettools as st
 import hf_hydrodata
 import pytest
+import glob
 
 
 _SECONDS_PER_HOUR = 3600
@@ -47,53 +48,107 @@ def mock_hf_paths(monkeypatch):
     monkeypatch.setattr(hf_hydrodata, "get_paths", mock_get_paths)
 
 
-def test_subset_static(tmp_path, mock_hf_data):
+def test_subset_static_filenames(tmp_path, mock_hf_data):
     test_dir = tmp_path / "test_static"
     test_dir.mkdir()
-    paths = st.subset_static(
+    var_list = ("var1", "var2", "var3")
+    static_paths = st.subset_static(
         ij_bounds=(0, 0, 10, 20),
-        dataset="my_dataset",
-        var_list=("var1", "var2"),
+        dataset="dataset",
+        var_list=var_list,
         write_dir=test_dir,
     )
-    assert os.path.isfile(paths["var1"])
-    data1 = read_pfb(paths["var1"])
-    assert np.array_equal(data1, np.ones((_DUMMY_NZ, 20, 10)))
-    assert os.path.isfile(paths["var2"])
-    data2 = read_pfb(paths["var2"])
-    assert np.array_equal(data2, np.ones((_DUMMY_NZ, 20, 10)))
+    expected_paths = {
+        var: glob.glob(os.path.join(test_dir, "*" + var + "*"), root_dir=test_dir)[0]
+        for var in var_list
+    }
+    assert static_paths == expected_paths
 
 
-@pytest.mark.parametrize(
-    "time_zone",
-    [
-        pytest.param("UTC", id="UTC timezone"),
-        pytest.param("EST", id="non-UTC timezone"),
-    ],
-)
-def test_subset_press_init_function(time_zone, tmp_path, mock_hf_data):
-    test_dir = tmp_path / "test_press_init"
+def test_subset_static_data(tmp_path, mock_hf_data):
+    test_dir = tmp_path / "test_static"
+    test_dir.mkdir()
+    static_paths = st.subset_static(
+        ij_bounds=(0, 0, 10, 20),
+        dataset="dataset",
+        var_list=("precipitation",),
+        write_dir=test_dir,
+    )
+    data = read_pfb(static_paths["precipitation"])
+    assert np.array_equal(data, np.ones((_DUMMY_NZ, 20, 10)))
+
+
+def test_subset_press_init_filenames(tmp_path, mock_hf_data):
+    test_dir = tmp_path / "test_press_filenames"
     test_dir.mkdir()
     file_path = st.subset_press_init(
         ij_bounds=(0, 0, 10, 20),
-        dataset="my_dataset",
+        dataset="dataset",
         date="2010-10-02",
         write_dir=test_dir,
-        time_zone=time_zone,
     )
-    assert os.path.isfile(file_path)
+    assert os.path.join(test_dir, os.listdir(test_dir)[0]) == file_path
+
+
+def test_subset_press_init_data(tmp_path, mock_hf_data):
+    test_dir = tmp_path / "test_press_data"
+    test_dir.mkdir()
+    file_path = st.subset_press_init(
+        ij_bounds=(0, 0, 10, 20),
+        dataset="dataset",
+        date="2010-10-02",
+        write_dir=test_dir,
+    )
     data = read_pfb(file_path)
     assert np.array_equal(data, np.ones((_DUMMY_NZ, 20, 10)))
 
 
 @pytest.mark.parametrize(
-    "time_zone",
+    "end_date, num_files",
     [
-        pytest.param("UTC", id="UTC timezone"),
[…]
