### 2024-07-10 #6752 @JannisHoch
Well to be honest, I am not really working with it and will likely not be working with it anytime soon. But I know that ParFlow, the model for which this utility software was developed, is widely used, mostly in the US but also outside.
I see application in research, but also for non-scientific purposes - if not even more for the latter than the former. AFAIK, the group developing ParFlow is relatively busy with non-academic outreach, as also indicated in the JOSS manuscript.
Data access went well. Regarding tests, I think they are aware now that for external collaborators they have to design tests differently to avoid internal dependencies.
