### 2024-07-10 #6752 @cheginit
@JannisHoch Thanks for the valuable feedback. I will take it from here. Appreciate your help and have a great rest of the day!

@gartavanis Please let me know if the issue with test has been addressed, and if not, what's your plan for addressing the issue? An important criterion for JOSS publication is its application, verifiability, and ease of use for external collaborator.
