### 2024-07-10 #6752 @dvalters
Hi @cheginit - I've moved on to a different institution now and no longer have access to the computer/environment I used to review the software originally. Sadly I won't have time to repeat the testing section to generate the error message as the installation process is quite lengthy (docker, registering for data access etc). I was happy that the software had sufficient tests, just that some of them appeared to require external datasets (but not all of them).
