### 2024-07-10 #6752 @gartavanis
Hi @cheginit, thanks for the comments. I will fix the references shortly.

The issues with the tests have been addressed in this issue created by @JannisHoch: https://github.com/hydroframe/subsettools/issues/86
