### 2024-07-10 05afa95
Test subsetting (#93)
* Monkeypatch hf.get_gridded_data for tests

* Add pseudocode for mock hf call

* Add more options to mock_get_data

* Fix bug in pfb dimension order

* adds subset press init test

* adds mocking and test for subset_forcing

* adds parameters to subset press init and subset forcing functions

* adds subset forcing file name testing

* Cleanup and format code

* Refactor tests

* Trigger build

* Bump version number

---------

Co-authored-by: gartavanis <gartavanis@users.noreply.github.com>
Co-authored-by: egwilborn <egwilborn@gmail.com>
