+++ tests/test_subsetting.py
@@ -2,6 +2,174 @@ import os
 from parflow.tools.io import read_pfb
 import numpy as np
 import subsettools as st
+import hf_hydrodata
+import pytest
+import glob
+
+
+_SECONDS_PER_HOUR = 3600
+_DUMMY_NZ = 5
+
+
+@pytest.fixture
+def mock_hf_data(monkeypatch):
+    def mock_get_data(options):
+        grid_bounds = options.get("grid_bounds")
+        resolution = options.get("temporal_resolution")
+        variable = options.get("variable")
+        start_time = options.get("start_time")
+        end_time = options.get("end_time")
+        imin, jmin, imax, jmax = grid_bounds
+        ni, nj = imax - imin, jmax - jmin
+        if resolution == "hourly":
+            if end_time is not None:
+                nt = int((end_time - start_time).total_seconds() / _SECONDS_PER_HOUR)
+            else:
+                nt = 1
+            if variable == "pressure_head":
+                shape = (nt, _DUMMY_NZ, nj, ni)
+            else:
+                shape = (nt, nj, ni)
+        elif resolution == "static":
+            shape = (_DUMMY_NZ, nj, ni)
+        return np.ones(shape)
+
+    monkeypatch.setattr(hf_hydrodata, "get_gridded_data", mock_get_data)
+
+
+@pytest.fixture
+def mock_hf_paths(monkeypatch):
+    def mock_get_paths(options):
+        dataset = options.get("dataset")
+        variable = options.get("variable")
+        file_path = [f"/folder/{dataset}.{variable}.000000_to_000000.pfb"]
+        return file_path
+
+    monkeypatch.setattr(hf_hydrodata, "get_paths", mock_get_paths)
+
+
+def test_subset_static_filenames(tmp_path, mock_hf_data):
+    test_dir = tmp_path / "test_static"
+    test_dir.mkdir()
+    var_list = ("var1", "var2", "var3")
+    static_paths = st.subset_static(
+        ij_bounds=(0, 0, 10, 20),
+        dataset="dataset",
+        var_list=var_list,
+        write_dir=test_dir,
+    )
+    expected_paths = {
+        var: glob.glob(os.path.join(test_dir, "*" + var + "*"), root_dir=test_dir)[0]
+        for var in var_list
+    }
+    assert static_paths == expected_paths
+
+
+def test_subset_static_data(tmp_path, mock_hf_data):
+    test_dir = tmp_path / "test_static"
+    test_dir.mkdir()
+    static_paths = st.subset_static(
+        ij_bounds=(0, 0, 10, 20),
+        dataset="dataset",
+        var_list=("precipitation",),
+        write_dir=test_dir,
+    )
+    data = read_pfb(static_paths["precipitation"])
+    assert np.array_equal(data, np.ones((_DUMMY_NZ, 20, 10)))
+
+
+def test_subset_press_init_filenames(tmp_path, mock_hf_data):
+    test_dir = tmp_path / "test_press_filenames"
+    test_dir.mkdir()
+    file_path = st.subset_press_init(
+        ij_bounds=(0, 0, 10, 20),
+        dataset="dataset",
+        date="2010-10-02",
+        write_dir=test_dir,
+    )
+    assert os.path.join(test_dir, os.listdir(test_dir)[0]) == file_path
+
+
+def test_subset_press_init_data(tmp_path, mock_hf_data):
+    test_dir = tmp_path / "test_press_data"
+    test_dir.mkdir()
+    file_path = st.subset_press_init(
+        ij_bounds=(0, 0, 10, 20),
+        dataset="dataset",
[…]
