+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "subsettools"
-version = "2.0.0"
+version = "2.0.1"
 description = "Subsetting tools and utilities for ParFlow"
 authors = ["George Artavanis, Amanda Triplett"]
 license = "MIT"
