### 2024-07-10 #6752 @gartavanis
@cheginit, I have created a new release of subsettools (https://github.com/hydroframe/subsettools/releases/tag/2.0.1). It's also on PyPI. This is version 2.0.1. The Zenodo archive is here: https://zenodo.org/records/12707059. (DOI: 10.5281/zenodo.12707059). Please let me know if there's anything else that remains to be done on our end. Thanks again for your time!
