### 2024-07-10 #6752 @cheginit
@gartavanis Thanks for working on addressing the issues diligently and congrats on making it this far! The EiC will take over the submission for the final publication, in the meanwhile, please check the final proof of the paper in the previous comment, just to make sure everything looks good such as authors' names.
