+++ tests/test_domain.py
@@ -168,27 +168,25 @@ def test_write_mask_solid(set_parflow_dir, tmp_path):
     )
     assert len(match) == 2 and len(mismatch) == 0 and len(errors) == 0
 
-@pytest.mark.parametrize(
-    "mask", [
-        pytest.param(np.array([[1]]),id="mask of just 1"),
-        pytest.param(np.array([[0,1],[0,1]]), id="simple 2D mask"),
-    ],
-)
-def test_write_mask_solid_simple(mask, set_parflow_dir,tmp_path):
-    test_dir = tmp_path / "test"
-    test_dir.mkdir()
-    st.write_mask_solid(mask=mask, grid="conus2", write_dir=test_dir)
-    assert os.path.isfile(f"{test_dir}/mask.pfb"), "Error generating mask.pfb file"
-    data = read_pfb(f'{test_dir}/mask.pfb')
-    assert (data == mask).all(), "Error: mask.pfb file generated does not match input array"
 
-def test_write_mask_solid_zero(set_parflow_dir, tmp_path):
-    mask = np.array([[0]])
-    result = np.array(['1\n', '0\n', '1\n', '0\n', '0\n'])
+@pytest.mark.parametrize(
+    "mask",
+    [np.array([[0]]), np.array([[1]]), np.array([[0, 1], [0, 1]])],
+)
+def test_write_mask(mask, set_parflow_dir, tmp_path):
     test_dir = tmp_path / "test"
     test_dir.mkdir()
     st.write_mask_solid(mask=mask, grid="conus2", write_dir=test_dir)
-    with open(f'{test_dir}/solidfile.pfsol') as f:
-        assert os.path.isfile(f'{test_dir}/solidfile.pfsol'), "Error generating solidfile.pfsol"
-        data = np.array(f.readlines())
-        assert (data == result).all(), "Error: solidfile.pfsol file generated does not match expected result array"
+    data = read_pfb(f"{test_dir}/mask.pfb")[0]
+    assert np.array_equal(data, mask)
+
+
+def test_write_solid(set_parflow_dir, tmp_path):
+    mask = np.array([[0]])
+    test_dir = tmp_path / "test"
+    test_dir.mkdir()
+    st.write_mask_solid(mask=mask, grid="conus2", write_dir=test_dir)
+    expected_solid = ['1\n', '0\n', '1\n', '0\n', '0\n']
+    with open(f"{test_dir}/solidfile.pfsol") as f:
+        solid  = f.readlines()
+    assert solid == expected_solid
