### 2024-07-10 d25d166
adds bug fix for mask value checking (#84)
* adds bug fix for mask value checking

* test commit from verde

* adds simple write_solid_mask tests

* adds mask dimension check

* adds file assertion and updates error messages

* updates file assertion path

* updates error message

* removes comment

* removes space

* removes allowed values variable

* uses with open statement

* Refactor tests

---------

Co-authored-by: Emily G Wilborn <ew7383@verde.princeton.edu>
Co-authored-by: gartavanis <gartavanis@users.noreply.github.com>
