### 2024-07-10 b36fc2d
adds testing conditions for define huc domain tests (#89)
* adds testing conditions for define huc domain tests

* adds testing conditions for define lat/lon domain tests

* corrects latlon domain test params
