+++ tests/test_domain.py
@@ -9,13 +9,15 @@ from parflow.tools.io import read_pfb
 @pytest.mark.parametrize(
     "hucs, grid, result",
     [
-        (["15060202"], "conus1", (375, 239, 487, 329)),
-        (["140802"], "conus1", (572, 337, 797, 577)),
-        (
+        pytest.param(["15060202"], "conus1", (375, 239, 487, 329),id="simple conus1 level 8 HUC domain"),
+        pytest.param(["140802"], "conus1", (572, 337, 797, 577), id="simple conus1 level 6 HUC domain"),
+        pytest.param(
             ["14080201", "14080202", "14080203", "14080204", "14080205"],
             "conus1",
-            (572, 337, 797, 577),
+            (572, 337, 797, 577), id="list of HUCs in conus 1"
         ),
+        pytest.param(["02050304"], "conus2", (3744, 1853, 3854, 1952), id="level 8 conus2 HUC domain"),
+        pytest.param(["17100306"], "conus2", (73, 2295, 110, 2368), id="level 8 conus2 coastal HUC domain")
     ],
 )
 def test_define_huc_domain(hucs, grid, result):
@@ -26,11 +28,14 @@ def test_define_huc_domain(hucs, grid, result):
 @pytest.mark.parametrize(
     "hucs, grid",
     [
-        (["01010001"], "conus1"),
-        (["01010001"], "conus2"),
-        (["03130003"], "conus1"),
-        (["1710"], "conus1"),
-        (["01010002", "01010001"], "conus1"),
+        pytest.param(["01010001"], "conus1", id="Level 8 HUC located outside of conus1 grid"),
+        pytest.param(["21010005"], "conus2", id="Level 8 HUC located outside of conus2 grid"),
+        pytest.param(["03130003"], "conus1", id="Level 8 HUC located outside of conus1 grid"),
+        pytest.param(["1710"], "conus1", id="Level 4 HUC located outside of conus1 grid"),
+        pytest.param(["01010002", "01010001"], "conus1", id="list of Level 8 HUCs located outside of conus1 grid"),
+        pytest.param(["220102"], "conus2", id="HUC number does not exist"),
+        pytest.param(["14080205", "140700"], "conus2", id="HUC IDs of different levels")
+
     ],
 )
 def test_define_huc_domain_errors(hucs, grid):
@@ -76,6 +81,8 @@ def test_define_latlon_domain(latlon_bounds, grid, correct_bounds, correct_mask)
             "conus2",
             id="lat/lon points outside the grid",
         ),
+        pytest.param([[57.44, -107.33]], "conus2", id="only one lat/lon point"),
+        pytest.param([[57.44, -107.33],[57.44, -107.33], [57.55, -108.00]], "conus2", id="more than two lat/lon points")
     ],
 )
 def test_define_latlon_domain_errors(latlon_bounds, grid):
