### 2024-07-15 a07aab9
adds filepath and data tests for get runscripts
