### 2024-07-17 #6752 @kthyng
Hi @gartavanis, I see that the version is `v2.0.1` here and the repo itself, but in Zenodo it is v1. Can you update that? While you're at it, I see that your title and author list in Zenodo already match the JOSS paper (thanks!), though the authors with middle initials included might appreciate them being in Zenodo also.
