### 2024-07-17 #6752 @gartavanis
Hi @kthyng, thanks for the comments! I added the version number on Zenodo and the middle initials.
