### 2024-07-17 98f9c09
Create CITATION.cff
