### 2024-07-17 #6752 @kthyng
Congratulations on your new publication @gartavanis! Many thanks to @cheginit and to reviewers @JannisHoch and @dvalters for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
